fix(menu_card): fix menu item card component
refactor component directory and fixed minor bugs
