--- src/app/api/url/route.ts
@@ -1,8 +1,0 @@
-import { getBaseUrl } from '@/utils/getBaseUrl'
-import { NextResponse } from 'next/server'
-
-export async function GET(
-) {
-  const baseUrl = await getBaseUrl()
-  return NextResponse.json({ baseUrl })
-}
