+++ src/app/dashboard/layout.tsx
@@ -1,12 +1,9 @@
-import { Inter } from 'next/font/google'
 import { redirect } from 'next/navigation'
 import { createClient } from '@/utils/supabase/server'
 import { SessionProvider } from '@/components/SessionProvider'
 import { Header } from '@/components/Header'
 import '@/app/globals.css'
 import { CustomToaster } from '@/components/CustomToaster'
-
-const inter = Inter({ subsets: ['latin'] })
 
 export default async function DashboardLayout({
   children,
