+++ src/components/Portal.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { useEffect, useState } from 'react'
+
 import { createPortal } from 'react-dom'
 
 interface PortalProps {
